Refactored position of most actions, added Blueprints for translate and edit
