+++ Apps.OpenAI/Actions/ReviewActions.cs
@@ -1,0 +1,173 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Apps.OpenAI.Actions.Base;
+using Apps.OpenAI.Constants;
+using Apps.OpenAI.Dtos;
+using Apps.OpenAI.Models.Identifiers;
+using Apps.OpenAI.Models.Requests.Chat;
+using Apps.OpenAI.Models.Responses.Chat;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Invocation;
+using Blackbird.Applications.SDK.Extensions.FileManagement.Interfaces;
+using Newtonsoft.Json;
+using MoreLinq;
+
+namespace Apps.OpenAI.Actions;
+
+[ActionList("Review")]
+public class ReviewActions(InvocationContext invocationContext, IFileManagementClient fileManagementClient)
+    : BaseActions(invocationContext, fileManagementClient)
+{
+    [Action("Get translation issues",
+        Description = "Review text translation and generate a comment with the issue description")]
+    public async Task<ChatResponse> GetTranslationIssues([ActionParameter] TextChatModelIdentifier modelIdentifier,
+        [ActionParameter] GetTranslationIssuesRequest input, [ActionParameter] GlossaryRequest glossary)
+    {
+        var systemPrompt =
+            $"You are receiving a source text{(input.SourceLanguage != null ? $" written in {input.SourceLanguage} " : "")}" +
+            $"that was translated by NMT into target text{(input.TargetLanguage != null ? $" written in {input.TargetLanguage}" : "")}. " +
+            "Evaluate the target text for grammatical errors, language structure issues, and overall linguistic coherence, " +
+            "including them in the issues description. Respond with the issues description. " +
+            $"{(input.TargetAudience != null ? $"The target audience is {input.TargetAudience}" : string.Empty)}";
+
+
+        if (glossary.Glossary != null)
+            systemPrompt +=
+                " Ensure that the translation aligns with the glossary entries provided for the respective " +
+                "languages, and note any discrepancies, ambiguities, or incorrect usage of terms. Include " +
+                "these observations in the issues description.";
+
+        if (input.AdditionalPrompt != null)
+            systemPrompt = $"{systemPrompt} {input.AdditionalPrompt}";
+
+        var userPrompt = @$"
+            Source text: 
+            {input.SourceText}
+
+            Target text: 
+            {input.TargetText}
+        ";
+
+        if (glossary.Glossary != null)
+        {
+            var glossaryPromptPart = await GetGlossaryPromptPart(glossary.Glossary, input.SourceText, true);
+            if (glossaryPromptPart != null) userPrompt += glossaryPromptPart;
+        }
+
+        var messages = new List<ChatMessageDto> { new(MessageRoles.System, systemPrompt), new(MessageRoles.User, userPrompt) };
+        var response = await ExecuteChatCompletion(messages, modelIdentifier.GetModel());
+
+        return new()
+        {
+            SystemPrompt = systemPrompt,
+            UserPrompt = userPrompt,
+            Message = response.Choices.First().Message.Content,
+            Usage = response.Usage,
+        };
+    }
+
+    [Action("Get MQM report",
+        Description = "Perform an LQA Analysis of the translation. The result will be in the MQM framework form.")]
+    public async Task<ChatResponse> GetLqaAnalysis([ActionParameter] TextChatModelIdentifier modelIdentifier,
+        [ActionParameter] GetTranslationIssuesRequest input, [ActionParameter] GlossaryRequest glossary)
+    {
+        var systemPrompt = "Perform an LQA analysis and use the MQM error typology format using all 7 dimensions. " +
+                           "Here is a brief description of the seven high-level error type dimensions: " +
+                           "1. Terminology – errors arising when a term does not conform to normative domain or organizational terminology standards or when a term in the target text is not the correct, normative equivalent of the corresponding term in the source text. " +
+                           "2. Accuracy – errors occurring when the target text does not accurately correspond to the propositional content of the source text, introduced by distorting, omitting, or adding to the message. " +
+                           "3. Linguistic conventions  – errors related to the linguistic well-formedness of the text, including problems with grammaticality, spelling, punctuation, and mechanical correctness. " +
+                           "4. Style – errors occurring in a text that are grammatically acceptable but are inappropriate because they deviate from organizational style guides or exhibit inappropriate language style. " +
+                           "5. Locale conventions – errors occurring when the translation product violates locale-specific content or formatting requirements for data elements. " +
+                           "6. Audience appropriateness – errors arising from the use of content in the translation product that is invalid or inappropriate for the target locale or target audience. " +
+                           "7. Design and markup – errors related to the physical design or presentation of a translation product, including character, paragraph, and UI element formatting and markup, integration of text with graphical elements, and overall page or window layout. " +
+                           "Provide a quality rating for each dimension from 0 (completely bad) to 10 (perfect). You are an expert linguist and your task is to perform a Language Quality Assessment on input sentences. " +
+                           "Try to propose a fixed translation that would have no LQA errors. " +
+                           "Formatting: use line spacing between each category. The category name should be bold.";
+
+        if (glossary.Glossary != null)
+            systemPrompt +=
+                " Use the provided glossary entries for the respective languages. If there are discrepancies " +
+                "between the translation and glossary, note them in the 'Terminology' part of the report, " +
+                "along with terminology problems not related to the glossary.";
+
+        if (input.AdditionalPrompt != null)
+            systemPrompt = $"{systemPrompt} {input.AdditionalPrompt}";
+
+        var userPrompt =
+            $"{(input.SourceLanguage != null ? $"The {input.SourceLanguage} " : "")}\"{input.SourceText}\" was translated as \"{input.TargetText}\"{(input.TargetLanguage != null ? $" into {input.TargetLanguage}" : "")}.{(input.TargetAudience != null ? $" The target audience is {input.TargetAudience}" : "")}";
+
+        if (glossary.Glossary != null)
+        {
+            var glossaryPromptPart = await GetGlossaryPromptPart(glossary.Glossary, input.SourceText, true);
+            if (glossaryPromptPart != null) userPrompt += glossaryPromptPart;
+        }
+
+        var messages = new List<ChatMessageDto> { new(MessageRoles.System, systemPrompt), new(MessageRoles.User, userPrompt) };
+        var response = await ExecuteChatCompletion(messages, modelIdentifier.GetModel());
+
+        return new()
+        {
+            SystemPrompt = systemPrompt,
+            UserPrompt = userPrompt,
+            Message = response.Choices.First().Message.Content,
+            Usage = response.Usage,
+        };
+    }
+
+    [Action("Get MQM dimension values",
+        Description =
+            "Perform an LQA Analysis of the translation. The result will be in the MQM framework form. This action " +
+            "only returns the scores (between 1 and 10) of each dimension.")]
+    public async Task<MqmAnalysis> GetLqaDimensionValues([ActionParameter] TextChatModelIdentifier modelIdentifier,
+        [ActionParameter] GetTranslationIssuesRequest input, [ActionParameter] GlossaryRequest glossary)
+    {
+        var systemPrompt = "Perform an LQA analysis and use the MQM error typology format using all 7 dimensions. " +
+                           "Here is a brief description of the seven high-level error type dimensions: " +
+                           "1. Terminology – errors arising when a term does not conform to normative domain or organizational terminology standards or when a term in the target text is not the correct, normative equivalent of the corresponding term in the source text. " +
+                           "2. Accuracy – errors occurring when the target text does not accurately correspond to the propositional content of the source text, introduced by distorting, omitting, or adding to the message. " +
+                           "3. Linguistic conventions  – errors related to the linguistic well-formedness of the text, including problems with grammaticality, spelling, punctuation, and mechanical correctness. " +
+                           "4. Style – errors occurring in a text that are grammatically acceptable but are inappropriate because they deviate from organizational style guides or exhibit inappropriate language style. " +
+                           "5. Locale conventions – errors occurring when the translation product violates locale-specific content or formatting requirements for data elements. " +
+                           "6. Audience appropriateness – errors arising from the use of content in the translation product that is invalid or inappropriate for the target locale or target audience. " +
+                           "7. Design and markup – errors related to the physical design or presentation of a translation product, including character, paragraph, and UI element formatting and markup, integration of text with graphical elements, and overall page or window layout. " +
+                           "Provide a quality rating for each dimension from 0 (completely bad) to 10 (perfect). You are an expert linguist and your task is to perform a Language Quality Assessment on input sentences. " +
+                           "Try to propose a fixed translation that would have no LQA errors. " +
+                           "The response should be in the following json format: " +
+                           "{\r\n  \"terminology\": 0,\r\n  \"accuracy\": 0,\r\n  \"linguistic_conventions\": 0,\r\n  \"style\": 0,\r\n  \"locale_conventions\": 0,\r\n  \"audience_appropriateness\": 0,\r\n  \"design_and_markup\": 0,\r\n  \"proposed_translation\": \"fixed translation\"\r\n}";
+
+        if (glossary.Glossary != null)
+            systemPrompt += " Use the provided glossary entries for the respective languages. If there are " +
+                            "discrepancies between the translation and glossary, reduce the score in the " +
+                            "'Terminology' part of the report respectively.";
+
+        if (input.AdditionalPrompt != null)
+            systemPrompt = $"{systemPrompt} {input.AdditionalPrompt}";
+
+        var userPrompt =
+            $"{(input.SourceLanguage != null ? $"The {input.SourceLanguage} " : "")}\"{input.SourceText}\" was translated as \"{input.TargetText}\"{(input.TargetLanguage != null ? $" into {input.TargetLanguage}" : "")}.{(input.TargetAudience != null ? $" The target audience is {input.TargetAudience}" : "")}";
+
+        if (glossary.Glossary != null)
+        {
+            var glossaryPromptPart = await GetGlossaryPromptPart(glossary.Glossary, input.SourceText, true);
+            if (glossaryPromptPart != null) userPrompt += glossaryPromptPart;
+        }
+
+        var messages = new List<ChatMessageDto> { new(MessageRoles.System, systemPrompt), new(MessageRoles.User, userPrompt) };
+        var response = await ExecuteChatCompletion(messages, modelIdentifier.GetModel(), input, new { type = "json_object" });
+
+        try
+        {
+            var analysis = JsonConvert.DeserializeObject<MqmAnalysis>(response.Choices.First().Message.Content);
+            analysis.Usage = response.Usage;
+            return analysis;
+        }
+        catch
+        {
+            throw new Exception(
+                "Something went wrong parsing the output from OpenAI, most likely due to a hallucination!");
+        }
+    }
+}
