+++ Apps.OpenAI/Models/Requests/Chat/LocalizeTextRequest.cs
@@ -2,13 +2,15 @@
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Dictionaries;
 using Blackbird.Applications.Sdk.Common.Dynamic;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.Translate;
 
 namespace Apps.OpenAI.Models.Requests.Chat;
 
-public class LocalizeTextRequest : BaseChatRequest
+public class LocalizeTextRequest : BaseChatRequest, ITranslateTextInput
 {
     public string Text { get; set; }
-        
+
+    [Display("Target language")]
     [StaticDataSource(typeof(LocaleDataSourceHandler))]
-    public string Locale { get; set; }
+    public string TargetLanguage { get; set; }
 }
