+++ Apps.OpenAI/Models/Responses/Chat/TranslateTextResponse.cs
@@ -1,10 +1,11 @@
 ﻿using Apps.OpenAI.Dtos;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.SDK.Blueprints.Interfaces.Edit;
+using Blackbird.Applications.SDK.Blueprints.Interfaces.Translate;
 
 namespace Apps.OpenAI.Models.Responses.Chat;
 
-public class EditResponse : IEditTextOutput
+public class TranslateTextResponse : ITranslateTextOutput
 {
     [Display("System prompt")]
     public string SystemPrompt { get; set; }
@@ -12,8 +13,8 @@ public class EditResponse : IEditTextOutput
     [Display("User prompt")]
     public string UserPrompt { get; set; }
 
-    [Display("Edited text")]
-    public string EditedText { get; set; }
+    [Display("Translated text")]
+    public string TranslatedText { get; set; }
 
     [Display("Usage")]
     public UsageDto Usage { get; set; }
